\name{rw.query.fields}
\alias{rw.query.fields}
\title{Reliefweb | Query Fields}
\usage{
rw.query.fields(field = NULL)
}
\arguments{
  \item{field}{provide a string of the field of interest.
  The results will contain all relevant fields.}
}
\description{
This function allows for the understanding of the query
fields. It gives a list of the fields, together with their
description and basic metadata about them.
}

